@@ -1,3 +1,7 @@
+/**
+ * @format
+ */
+
 import 'react-native';
 import React from 'react';
 import Root from '../src/index';
@@ -6,7 +10,5 @@ import Root from '../src/index';
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  const tree = renderer.create(
-    <Root />
-  );
+  renderer.create(<Root />);
 });
