@@ -1,4 +1,9 @@
-import { AppRegistry } from 'react-native';
-import Root from './src/index';
+/**
+ * @format
+ */
 
-AppRegistry.registerComponent('example', () => Root);
+import {AppRegistry} from 'react-native';
+import App from './src';
+import {name as appName} from './app.json';
+
+AppRegistry.registerComponent(appName, () => App);
