@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
 import { ParallaxImage } from 'react-native-snap-carousel';
-import styles from 'example/src/styles/SliderEntry.style';
+import styles from '../styles/SliderEntry.style';
 
 export default class SliderEntry extends Component {
 
