@@ -26,7 +26,11 @@ export default class ParallaxImage extends Component {
         fadeDuration: PropTypes.number,
         parallaxFactor: PropTypes.number,
         showSpinner: PropTypes.bool,
-        spinnerColor: PropTypes.string
+        spinnerColor: PropTypes.string,
+        AnimatedImageComponent: PropTypes.oneOfType([
+            PropTypes.func,
+            PropTypes.object
+        ])
     };
 
     static defaultProps = {
@@ -34,7 +38,8 @@ export default class ParallaxImage extends Component {
         fadeDuration: 500,
         parallaxFactor: 0.3,
         showSpinner: true,
-        spinnerColor: 'rgba(0, 0, 0, 0.4)'
+        spinnerColor: 'rgba(0, 0, 0, 0.4)',
+        AnimatedImageComponent: Animated.Image
     }
 
     constructor(props) {
@@ -150,6 +155,7 @@ export default class ParallaxImage extends Component {
             parallaxFactor,
             style,
             source,
+            AnimatedImageComponent,
             ...other
         } = this.props;
 
@@ -188,12 +194,11 @@ export default class ParallaxImage extends Component {
         }
 
         return (
-            <Animated.Image
-                {...other}
-                source={source}
-                style={[styles.image, style, requiredStyles, dynamicStyles]}
-                onLoad={this._onLoad}
-                onError={status !== 3 ? this._onError : undefined} // prevent infinite-loop bug
+            <AnimatedImageComponent
+              {...other}
+              style={[styles.image, style, requiredStyles, dynamicStyles]}
+              onLoad={this._onLoad}
+              onError={status !== 3 ? this._onError : undefined} // prevent infinite-loop bug
             />
         );
     }
